+++ libs/ngx-slice-kit/src/lib/buttons/button/button.component.ts
@@ -6,7 +6,7 @@ const AVAILABLE_SIZE_CLASS_NAMES = [
 ];
 
 const BUTTON_CLASS_NAMES = [
-    'sdk-button',
+    'sdk-base-button',
     'sdk-flat-button',
     'sdk-raised-button',
     'sdk-stroked-button',
@@ -15,7 +15,7 @@ const BUTTON_CLASS_NAMES = [
 
 @Component({
     // tslint:disable-next-line:component-selector
-    selector: '[sdk-button], [sdk-flat-button], [sdk-raised-button], [sdk-stroked-button], [sdk-round-button]',
+    selector: '[sdk-base-button], [sdk-flat-button], [sdk-raised-button], [sdk-stroked-button], [sdk-round-button]',
     // templateUrl: './button.component.html',
     template: `
         <ng-content></ng-content>