+++ libs/ngx-slice-kit/src/lib/core/theme/lib/theme-light.ts
@@ -3,6 +3,7 @@ import { defaultColors, Theme } from '../theme.model';
 export const themeLight: Theme = {
     name: 'light',
     base: defaultColors.base,
+    base_text: defaultColors.background,
     background: defaultColors.background,
     regular: defaultColors.regular,
     regular_text: defaultColors.base,