+++ libs/ngx-slice-kit/src/lib/core/theme/theme.directive.ts
@@ -58,6 +58,10 @@ export class ThemeDirective implements OnInit, OnDestroy {
                 styles.innerHTML += (`${prop.prop}-rgb: ${prop.rgb};`);
                 styles.innerHTML += (`${prop.prop}-opposite: ${prop.background};`);
                 styles.innerHTML += (`${prop.prop}: ${prop.hex};`);
+                if (prop.text_value) {
+                    styles.innerHTML += (`${prop.prop}-text-raw: ${prop.text_value};`);
+                    styles.innerHTML += (`${prop.prop}-text: ${prop.text};`);
+                }
             }
         }
         styles.innerHTML += `}`;