+++ libs/ngx-slice-kit/src/lib/core/theme/theme.model.ts
@@ -31,36 +31,15 @@ export const colors = [
     `regular`, `primary`, `success`, `accent`, `warn`,
 ];
 
-export const rgbDepth = `rgb`;
-export const emptyDepth = ``;
-
-class Depth {
-    color: string;
-    alpha: number = 1;
-    background?: string;
-}
-
-export const colorsDepth = {
-    rgb: {alpha: 100},
-    deep: {alpha: 90},
-    hover: {alpha: 75},
-    active: {alpha: 60},
-    disabled: {alpha: 45},
-    smooth: {alpha: 10},
-};
-
-export const textDepth = {
-    placeholder: {color: 'background', alpha: 75, background: 'base'},
-    text: {color: 'background', alpha: 5, background: 'base'},
-};
-
 const alphaLimit = 100;
 const alphaStep = 10;
 
 export class ColorProperty {
     value: string;
+    text_value?: string;
     background: string;
     name: string;
+    text?: string;
     prop?: string;
     alpha?: number;
     rgb?: string;
@@ -68,13 +47,25 @@ export class ColorProperty {
 
     constructor(cp?: ColorProperty) {
         Object.assign(this, cp);
-
         this.prop = `--${this.name}`;
+
+        // find color values
         const rgbVal = this.value.split(sep(this.value)).map(i => parseInt(i, 10));
         const bgVal = this.background.split(sep(this.background)).map(i => parseInt(i, 10));
-        this.rgb = rgbaToRgb(rgbVal[0], rgbVal[1], rgbVal[2], this.alpha / 100, bgVal[0], bgVal[1], bgVal[2]);
+        const rgbMixin = mixinRgba(rgbVal, bgVal, this.alpha / 100);
+        const bgMixin = mixinRgba(bgVal, rgbVal, this.alpha / 100);
+
+        // declare rgb values
+        this.rgb = numArrayToRgbString(rgbMixin);
         this.hex = rgbaToHex(this.rgb);
-        this.background = rgbaToRgb(bgVal[0], bgVal[1], bgVal[2], this.alpha / 100, rgbVal[0], rgbVal[1], rgbVal[2]);
+        this.background = numArrayToRgbString(bgMixin);
+
+        // check if text value is available
+        if (this.text_value?.length > 0) {
+            const textVal = this.text_value.split(sep(this.text_value)).map(i => parseInt(i, 10));
+            const textMixin = mixinRgba(textVal, bgVal, 0.9);
+            this.text = numArrayToRgbString(textMixin);
+        }
     }
 }
 
@@ -90,6 +81,7 @@ export class Theme {
     accent?: string;
     warn?: string;
     // text keys
+    base_text?: string;
     regular_text?: string;
     primary_text?: string;
     success_text?: string;
@@ -106,8 +98,10 @@ export class Theme {
         for (const color of [...baseColors, ...colors]) {
             const rgb = this[color];
 
+            const textColor = `${color}_text`;
             const colorProp = {
                 name: color,
+                text_value: this[textColor],
                 alpha: 100,
                 value: rgb,
                 background: rgb === this.background ? this.base : this.background,
@@ -127,6 +121,25 @@ export class Theme {
         return results;
     }
 }
+
+const sep = (str: string): string => {
+    return str?.indexOf(',') < 0 ? (str?.indexOf(', ') < 0 ? ' ' : ', ') : ',';
+};
+
+const trim = (str: string): string => {
+    return str?.replace(/^\s+|\s+$/gm, '');
+};
+
+export const mixinRgba = (base: number[], added: number[], alpha: number): number[] => {
+    const r3 = Math.round(((1 - alpha) * added[0]) + (alpha * base[0]));
+    const g3 = Math.round(((1 - alpha) * added[1]) + (alpha * base[1]));
+    const b3 = Math.round(((1 - alpha) * added[2]) + (alpha * base[2]));
+    return [r3, g3, b3];
+};
+
+export const numArrayToRgbString = (rgb: number[]): string => {
+    return `rgb(${rgb[0]},${rgb[1]},${rgb[2]})`;
+};
 
 export const rgbaToHex = (rgba: string): string => {
     if (!rgba) {
@@ -158,28 +171,3 @@ export const rgbaToHex = (rgba: string): string => {
     return ('#' + outParts.join(''));
 };
 
-const sep = (str: string): string => {
-    return str?.indexOf(',') < 0 ? (str?.indexOf(', ') < 0 ? ' ' : ', ') : ',';
-};
-
-const trim = (str: string): string => {
-    return str?.replace(/^\s+|\s+$/gm, '');
-};
-
-export const mixinRgba = (base: number[], added: number[]): string => {
-    const mix = [];
-
-    mix[3] = 1 - (1 - added[3]) * (1 - base[3]); // alpha
-    mix[0] = Math.round((added[0] * added[3] / mix[3]) + (base[0] * base[3] * (1 - added[3]) / mix[3])); // red
-    mix[1] = Math.round((added[1] * added[3] / mix[3]) + (base[1] * base[3] * (1 - added[3]) / mix[3])); // green
-    mix[2] = Math.round((added[2] * added[3] / mix[3]) + (base[2] * base[3] * (1 - added[3]) / mix[3])); // blue
-
-    return `rgba(${base.join(',')})`;
-};
-
-export const rgbaToRgb = (r, g, b, a, r2, g2, b2: number): string => {
-    const r3 = Math.round(((1 - a) * r2) + (a * r));
-    const g3 = Math.round(((1 - a) * g2) + (a * g));
-    const b3 = Math.round(((1 - a) * b2) + (a * b));
-    return `rgb(${r3},${g3},${b3})`;
-};